requirements.txt
# Testing
mock== 1.3.0
httpretty==0.8.10
django-jenkins==0.18.1
pyflakes==1.0.0
pep8==1.6.2
freezegun==0.3.5
coverage==4.0.3

# Project Base
# 1.8 is LTS
Django==1.8.7 # rq.filter: >=1.8.7,< 1.9
#South==1.0
MySQL-python==1.2.5
premailer==2.9.6
#-e git+git@github.com:ArabellaTech/premailer@fix-external-path#egg=premailer
django-taggit==0.17.5 # for cms
django-basic-cms==0.3.11
django-compressor==1.6
django-mptt==0.7.4
pytz>2012
contactBox==0.1.5
Pillow==3.0.0
django-tinymce==2.0.6
django-grappelli==2.7.2
django-filebrowser==3.6.1
raven==5.1.1 # rq.filter: >=5.0.0,< 5.2.0  # sentry update required - painfull.
#django-modeltranslation==0.4.1
ydcommon==0.1.42
lxml==3.5.0
BeautifulSoup==3.2.1
flup==1.0.2
#celery==3.1.18
#django-celery==3.1.16
#amqp==1.4.6
#librabbitmq==1.6.1
django-bower==5.0.4
#djangorestframework==3.1.2
#django-rest-swagger==0.3.0
django-filter==0.11.0
django-templateaddons==0.1
django-su==0.5.0
#git+https://github.com/simul14/django-fields.git#egg=django-fields
sorl-thumbnail==12.3
django-passwords==0.3.4
django-storages-redux==1.3
boto==2.38.0
djangorestframework==3.3.1
django-rest-swagger==0.3.4
